Some u_long -> unsigned int changes.
These are valid because the invention and major uses of u_long in this
codebase predate the 64-bit transition - it may look like we're
narrowing size from 64 to 32 bits, but what we're actually doing is
replacing an earlier hack for guaranteeing 32-bit size by people
who cut their teeth on 16-bit ints.

We might actually get a little optimization out of this, but the
real reason is to get rid of the obfuscatory typedef.

+++ include/ntp_stdlib.h
@@ -176,13 +176,13 @@ extern	void	rereadkeys	(void);
  */
 
 /* authkeys.c */
-extern unsigned long	authkeynotfound;	/* keys not found */
-extern unsigned long	authkeylookups;		/* calls to lookup keys */
-extern unsigned long	authnumkeys;		/* number of active keys */
-extern unsigned long	authkeyexpired;		/* key lifetime expirations */
-extern unsigned long	authkeyuncached;	/* cache misses */
-extern unsigned long	authencryptions;	/* calls to encrypt */
-extern unsigned long	authdecryptions;	/* calls to decrypt */
+extern unsigned int	authkeynotfound;	/* keys not found */
+extern unsigned int	authkeylookups;		/* calls to lookup keys */
+extern unsigned int	authnumkeys;		/* number of active keys */
+extern unsigned int	authkeyexpired;		/* key lifetime expirations */
+extern unsigned int	authkeyuncached;	/* cache misses */
+extern unsigned int	authencryptions;	/* calls to encrypt */
+extern unsigned int	authdecryptions;	/* calls to decrypt */
 
 extern int	authnumfreekeys;
